pyproject.toml
[project]
name = "np_config"

[tool.poetry]
name = "np_config"
version = "0.1.3"
description = "Zookeeper configs with local backup, repackaging code from AIBS mpeconfig."
authors = [
    "Ben Hardcastle <ben.hardcastle@alleninstitute.org>",
    "Ross Hytnen <rossh@alleninstitute.org>", 
    "Ben Sutton <ben.sutton@alleninstitute.org>",
]
maintainers = ["Ben Hardcastle <ben.hardcastle@alleninstitute.org>"]
readme = "README.md"
include = ["resources/*"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[tool.poetry.urls]
"Repository" = "https://github.com/alleninstitute/np_config"
"Bug Tracker" = "https://github.com/alleninstitute/np_config/issues"

[tool.poetry.dependencies]
python = "^3.6"
kazoo = "~2.8"
pyyaml = "^5.3.0"

[tool.poetry.dev-dependencies]
pytest = "^7"
black = "*"
mypy = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]